Inviting mr imo to project

+++ lib/ui/screens/dashboard/fragments/report/model/upgrade_sign_up.dart
@@ -19,7 +19,7 @@ class UpgradeSignUpModel {
   String? year;
   String? month;
   int? signups;
-  int? income;
+  double? income;
 
   UpgradeSignUpModel({this.year, this.month, this.signups, this.income});
